requirements.txt
python-dotenv
openai
nfl_data_py
pandas
rich
